Types

None of the types below need to be constructed directly; they arise from performing operations between LinearMap objects or by calling the LinearMap constructor described in Methods.

LinearMap

Abstract supertype

LinearMaps.LinearMap — Type

LinearMap(A; kwargs...)
LinearMap(J, M::Int)
LinearMap{T=Float64}(f, [fc,], M::Int, N::Int = M; kwargs...)

Construct a linear map object, either from an existing LinearMap or AbstractMatrix A, with the purpose of redefining its properties via the keyword arguments kwargs; a UniformScaling object J with specified (square) dimension M; or from a function or callable object f. In the latter case, one also needs to specify the size of the equivalent matrix representation (M, N), i.e., for functions f acting on length N vectors and producing length M vectors (with default value N=M). Preferably, also the eltype T of the corresponding matrix representation needs to be specified, i.e. whether the action of f on a vector will be similar to, e.g., multiplying by numbers of type T. If not specified, the devault value T=Float64 will be assumed. Optionally, a corresponding function fc can be specified that implements the transpose/adjoint of f.

The keyword arguments and their default values for functions f are

issymmetric::Bool = false : whether A or f acts as a symmetric matrix

ishermitian::Bool = issymmetric & T<:Real : whether A or f acts as a Hermitian matrix

isposdef::Bool = false : whether A or f acts as a positive definite matrix.

For existing linear maps or matrices A, the default values will be taken by calling issymmetric, ishermitian and isposdef on the existing object A.

For functions f, there is one more keyword argument

ismutating::Bool : flags whether the function acts as a mutating matrix multiplication f(y,x) where the result vector y is the first argument (in case of true), or as a normal matrix multiplication that is called as y=f(x) (in case of false). The default value is guessed by looking at the number of arguments of the first occurrence of f in the method table.

General purpose method to construct LinearMap objects of specific types, as described in the Types section below

LinearMap{T}(A::AbstractMatrix[; issymmetric::Bool, ishermitian::Bool, isposdef::Bool])
LinearMap{T}(A::LinearMap [; issymmetric::Bool, ishermitian::Bool, isposdef::Bool])

Create a WrappedMap object that will respond to the methods issymmetric, ishermitian, isposdef with the values provided by the keyword arguments, and to eltype with the value T. The default values correspond to the result of calling these methods on the argument A; in particular {T} does not need to be specified and is set as eltype(A). This allows to use an AbstractMatrix within the LinearMap framework and to redefine the properties of an existing LinearMap.

LinearMap{T}(f, [fc = nothing], M::Int, [N::Int = M]; ismutating::Bool, issymmetric::Bool, ishermitian::Bool, isposdef::Bool])

Create a FunctionMap instance that wraps an object describing the action of the linear map on a vector as a function call. Here, f can be a function or any other object for which either the call f(src::AbstractVector) -> dest::AbstractVector (when ismutating = false) or f(dest::AbstractVector,src::AbstractVector) -> dest (when ismutating = true) is supported. The value of ismutating can be specified, by default its value is guessed by looking at the number of arguments of the first method in the method list of f.

A second function or object fc can optionally be provided that implements the action of the adjoint (transposed) linear map. Here, it is always assumed that this represents the adjoint/conjugate transpose, though this is of course equivalent to the normal transpose for real linear maps. Furthermore, the conjugate transpose also enables the use of mul!(y, transpose(A), x) using some extra conjugation calls on the input and output vector. If no second function is provided, than mul!(y, transpose(A), x) and mul!(y, adjoint(A), x) cannot be used with this linear map, unless it is symmetric or hermitian.

M is the number of rows (length of the output vectors) and N the number of columns (length of the input vectors). When the latter is not specified, N = M.

Finally, one can specify the eltype of the resulting linear map using the type parameter T. If not specified, a default value of Float64 is assumed. Use a complex type T if the function represents a complex linear map.

The keyword arguments and their default values are:

ismutating: false if the function f accepts a single vector argument corresponding to the input, and true if it accepts two vector arguments where the first will be mutated so as to contain the result. In both cases, the resulting A::FunctionMap will support both the mutating and non-mutating matrix-vector multiplication. Default value is guessed based on the number of arguments for the first method in the method list of f; it is not possible to use f and fc where only one of the two is mutating and the other is not.

issymmetric [=false]: whether the function represents the multiplication with a symmetric matrix. If true, this will automatically enable A' * x and transpose(A) * x.

ishermitian [=T<:Real && issymmetric]: whether the function represents the multiplication with a hermitian matrix. If true, this will automatically enable A' * x and transpose(A) * x.

isposdef [=false]: whether the function represents the multiplication with a positive definite matrix.

LinearMap(J::UniformScaling, M::Int)

Create a UniformScalingMap instance that corresponds to a uniform scaling map of size M×M.

FunctionMap

Type for wrapping an arbitrary function that is supposed to implement the matrix-vector product as a LinearMap.

WrappedMap

Type for wrapping an AbstractMatrix or LinearMap and to possible redefine the properties isreal, issymmetric, ishermitian and isposdef. An AbstractMatrix will automatically be converted to a WrappedMap when it is combined with other AbstractLinearMap objects via linear combination or composition (multiplication). Note that WrappedMap(mat1)*WrappedMap(mat2) will never evaluate mat1*mat2, since this is more costly than evaluating mat1*(mat2*x) and the latter is the only operation that needs to be performed by LinearMap objects anyway. While the cost of matrix addition is comparable to matrix-vector multiplication, this too is not performed explicitly since this would require new storage of the same amount as of the original matrices.

ScaledMap

Type for representing a scalar multiple of any LinearMap type. A ScaledMap will be automatically constructed if real or complex LinearMap objects are multiplied by real or complex scalars from the left or from the right.

UniformScalingMap

Type for representing a scalar multiple of the identity map (a.k.a. uniform scaling) of a certain size M=N, obtained simply as UniformScalingMap(λ, M). The type T of the resulting LinearMap object is inferred from the type of λ. A UniformScalingMap of the correct size will be automatically constructed if LinearMap objects are multiplied by scalars from the left or from the right (respecting the order of multiplication), if either the eltype of the LinearMap or the scalar are of non-commutative type, .

LinearCombination. CompositeMap, TransposeMap and AdjointMap

Used to add and multiply LinearMap objects, don't need to be constructed explicitly.

Base.:+ — Method

+(A::LinearMap, B::LinearMap)

Construct a LinearCombination <: LinearMap, a (lazy) representation of the sum of the two operators. Sums of LinearMap/LinearCombination objects and LinearMap/LinearCombination objects are reduced to a single LinearCombination. In sums of LinearMaps and AbstractMatrix/UniformScaling objects, the latter get promoted to LinearMaps automatically.

Examples

julia> CS = LinearMap{Int}(cumsum, 3)::LinearMaps.FunctionMap;

julia> LinearMap(ones(Int, 3, 3)) + CS + I + rand(3, 3);

Base.:* — Method

*(A::LinearMap, B::LinearMap)

Construct a CompositeMap <: LinearMap, a (lazy) representation of the product of the two operators. Products of LinearMap/CompositeMap objects and LinearMap/CompositeMap objects are reduced to a single CompositeMap. In products of LinearMaps and AbstractMatrix/UniformScaling objects, the latter get promoted to LinearMaps automatically.

Examples

julia> CS = LinearMap{Int}(cumsum, 3)::LinearMaps.FunctionMap;

julia> LinearMap(ones(Int, 3, 3)) * CS * I * rand(3, 3);

Missing docstring.

Missing docstring for LinearAlgebra.transpose(::LinearMap). Check Documenter's build log for details.

Missing docstring.

Missing docstring for LinearAlgebra.adjoint(::LinearMap). Check Documenter's build log for details.

KroneckerMap and KroneckerSumMap

Types for representing Kronecker products and Kronecker sums, resp., lazily.

Base.kron — Method

kron(A::LinearMap, B::LinearMap)
kron(A, B, Cs...)

Construct a KroneckerMap <: LinearMap object, a (lazy) representation of the Kronecker product of two LinearMaps. One of the two factors can be an AbstractMatrix, which is then promoted to a LinearMap automatically.

To avoid fallback to the generic Base.kron in the multi-map case, there must be a LinearMap object among the first 8 arguments in usage like kron(A, B, Cs...).

For convenience, one can also use A ⊗ B or ⊗(A, B, Cs...) (typed as \otimes+TAB) to construct the KroneckerMap, even when all arguments are of AbstractMatrix type.

If A, B, C and D are linear maps of such size that one can form the matrix products A*C and B*D, then the mixed-product property (A⊗B)*(C⊗D) = (A*C)⊗(B*D) holds. Upon vector multiplication, this rule is checked for applicability.

Examples

julia> J = LinearMap(I, 2) # 2×2 identity map
LinearMaps.UniformScalingMap{Bool}(true, 2)

julia> E = spdiagm(-1 => trues(1)); D = E + E' - 2I;

julia> Δ = kron(D, J) + kron(J, D); # discrete 2D-Laplace operator

julia> Matrix(Δ)
4×4 Array{Int64,2}:
 -4   1   1   0
  1  -4   0   1
  1   0  -4   1
  0   1   1  -4

LinearMaps.:⊗ — Function

⊗(k::Integer)

Construct a lazy representation of the k-th Kronecker power A^⊗(k) = A ⊗ A ⊗ ... ⊗ A, where A can be an AbstractMatrix or a LinearMap.

LinearMaps.kronsum — Function

kronsum(A, B)
kronsum(A, B, Cs...)

Construct a KroneckerSumMap <: LinearMap object, a (lazy) representation of the Kronecker sum A⊕B = A ⊗ Ib + Ia ⊗ B of two square linear maps of type LinearMap or AbstractMatrix. Here, Ia and Ib are identity operators of the size of A and B, respectively. Arguments of type AbstractMatrix are automatically promoted to LinearMaps.

For convenience, one can also use A ⊕ B or ⊕(A, B, Cs...) (typed as \oplus+TAB) to construct the KroneckerSumMap.

Examples

julia> J = LinearMap(I, 2) # 2×2 identity map
LinearMaps.UniformScalingMap{Bool}(true, 2)

julia> E = spdiagm(-1 => trues(1)); D = LinearMap(E + E' - 2I);

julia> Δ₁ = kron(D, J) + kron(J, D); # discrete 2D-Laplace operator, Kronecker sum

julia> Δ₂ = kronsum(D, D);

julia> Δ₃ = D^⊕(2);

julia> Matrix(Δ₁) == Matrix(Δ₂) == Matrix(Δ₃)
true

LinearMaps.:⊕ — Function

⊕(k::Integer)

Construct a lazy representation of the k-th Kronecker sum power A^⊕(k) = A ⊕ A ⊕ ... ⊕ A, where A can be a square AbstractMatrix or a LinearMap.

BlockMap and BlockDiagonalMap

Types for representing block (diagonal) maps lazily.

Base.hcat — Function

hcat(As::Union{LinearMap,UniformScaling}...)

Construct a BlockMap <: LinearMap object, a (lazy) representation of the horizontal concatenation of the arguments. UniformScaling objects are promoted to LinearMap automatically. To avoid fallback to the generic Base.hcat, there must be a LinearMap object among the first 8 arguments.

Examples

julia> CS = LinearMap{Int}(cumsum, 3)::LinearMaps.FunctionMap;

julia> L = [CS LinearMap(ones(Int, 3, 3))]::LinearMaps.BlockMap;

julia> L * ones(Int, 6)
3-element Array{Int64,1}:
 4
 5
 6

Base.vcat — Function

vcat(As::Union{LinearMap,UniformScaling}...)

Construct a BlockMap <: LinearMap object, a (lazy) representation of the vertical concatenation of the arguments. UniformScaling objects are promoted to LinearMap automatically. To avoid fallback to the generic Base.vcat, there must be a LinearMap object among the first 8 arguments.

Examples

julia> CS = LinearMap{Int}(cumsum, 3)::LinearMaps.FunctionMap;

julia> L = [CS; LinearMap(ones(Int, 3, 3))]::LinearMaps.BlockMap;

julia> L * ones(Int, 3)
6-element Array{Int64,1}:
 1
 2
 3
 3
 3
 3

Base.hvcat — Function

hvcat(rows::Tuple{Vararg{Int}}, As::Union{LinearMap,UniformScaling}...)

Construct a BlockMap <: LinearMap object, a (lazy) representation of the horizontal-vertical concatenation of the arguments. The first argument specifies the number of arguments to concatenate in each block row. UniformScaling objects are promoted to LinearMap automatically. To avoid fallback to the generic Base.hvcat, there must be a LinearMap object among the first 8 arguments.

Examples

julia> CS = LinearMap{Int}(cumsum, 3)::LinearMaps.FunctionMap;

julia> L = [CS CS; CS CS]::LinearMaps.BlockMap;

julia> L.rows
(2, 2)

julia> L * ones(Int, 6)
6-element Array{Int64,1}:
 2
 4
 6
 2
 4
 6